requirements.txt
Flask-SQLAlchemy==2.2